+++ rpc/types/types.go
@@ -16,6 +16,7 @@
 package types
 
 import (
+	"encoding/json"
 	"fmt"
 	"math/big"
 
@@ -23,6 +24,7 @@ import (
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/evmos/ethermint/x/evm/statedb"
+	evmtypes "github.com/evmos/ethermint/x/evm/types"
 )
 
 // Copied the Account and StorageResult types since they are registered under an
@@ -138,4 +140,9 @@ type OneFeeHistory struct {
 	BaseFee, NextBaseFee *big.Int   // base fee for each block
 	Reward               []*big.Int // each element of the array will have the tip provided to miners for the percentile given
 	GasUsedRatio         float64    // the ratio of gas used to the gas limit for each block
+}
+
+type TraceConfig struct {
+	evmtypes.TraceConfig
+	TracerConfig json.RawMessage `json:"tracerConfig"`
 }